requirements.txt
fastapi
uvicorn[standard]
semantic-kernel
openai
faiss-cpu
langsmith
langchain-openai
python-dotenv
pandas
streamlit
requests
sqlalchemy
jinja2
python-multipart
starlette
itsdangerous
aiofiles
httpx
